requirements.txt
requests
google-cloud-secretmanager
boxsdk
pyyaml
google-auth
google-cloud-firestore
